# openapi.yaml
openapi: 3.0.3

info:
  title: Treuepunkte Service API
  version: 1.0.0
  description: |
    Minimal OpenAPI specification for the Treuepunkte Service.
    Covers the currently implemented health, points, balance, and transaction endpoints.

servers:
  - url: http://localhost:8080
    description: Local Docker environment
  - url: https://{apiId}.execute-api.eu-west-1.amazonaws.com
    description: AWS API Gateway
    variables:
      apiId:
        default: example-id

tags:
  - name: Health
  - name: Points
  - name: Customers

paths:
  /health:
    get:
      tags:
        - Health
      summary: Health check
      operationId: getHealth
      responses:
        '200':
          description: Service is healthy
          content:
            text/plain:
              schema:
                type: string
                example: ok

  /v1/points/accrue:
    post:
      tags:
        - Points
      summary: Create pending loyalty points
      operationId: postAccrue
      parameters:
        - $ref: '#/components/parameters/IdempotencyKeyHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccrueRequest'
            examples:
              sample:
                value:
                  customer_id: cust-1
                  order_id: order-1001
                  home24_merch_cents: 1200
                  mirakl_merch_cents: 1000
                  shipping_cents: 490
                  currency: EUR
                  idempotency_key: accrue-1001
      responses:
        '201':
          description: Accrual created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
              examples:
                created:
                  value:
                    status: created
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /v1/points/confirm:
    post:
      tags:
        - Points
      summary: Confirm previously accrued points
      operationId: postConfirm
      parameters:
        - $ref: '#/components/parameters/IdempotencyKeyHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConfirmRequest'
            examples:
              sample:
                value:
                  customer_id: cust-1
                  order_id: order-1001
                  idempotency_key: confirm-1001
      responses:
        '201':
          description: Confirmation created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
              examples:
                created:
                  value:
                    status: created
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          description: Accrue transaction not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                accrueTransactionNotFound:
                  value:
                    error: accrue transaction not found
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /v1/points/revoke:
    post:
      tags:
        - Points
      summary: Revoke points for a returned order
      operationId: postRevoke
      parameters:
        - $ref: '#/components/parameters/IdempotencyKeyHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RevokeRequest'
            examples:
              sample:
                value:
                  customer_id: cust-1
                  order_id: order-1001
                  return_id: return-1001
                  idempotency_key: revoke-1001
      responses:
        '201':
          description: Revoke created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
              examples:
                created:
                  value:
                    status: created
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          description: Accrue transaction not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                accrueTransactionNotFound:
                  value:
                    error: accrue transaction not found
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /v1/points/redeem:
    post:
      tags:
        - Points
      summary: Redeem active loyalty points
      operationId: postRedeem
      parameters:
        - $ref: '#/components/parameters/IdempotencyKeyHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RedeemRequest'
            examples:
              sample:
                value:
                  customer_id: cust-1
                  reference: redeem-1001
                  points: 50
                  idempotency_key: redeem-1001
      responses:
        '201':
          description: Redemption created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
              examples:
                created:
                  value:
                    status: created
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /v1/points/restore:
    post:
      tags:
        - Points
      summary: Restore previously redeemed points
      operationId: postRestore
      parameters:
        - $ref: '#/components/parameters/IdempotencyKeyHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RestoreRequest'
            examples:
              sample:
                value:
                  customer_id: cust-1
                  reference: redeem-1001
                  idempotency_key: restore-1001
      responses:
        '201':
          description: Restore created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
              examples:
                created:
                  value:
                    status: created
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          description: Redeem transaction not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                redeemTransactionNotFound:
                  value:
                    error: redeem transaction not found
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /v1/customers/{customerId}/balance:
    get:
      tags:
        - Customers
      summary: Get customer balance
      operationId: getCustomerBalance
      parameters:
        - $ref: '#/components/parameters/CustomerIdPath'
      responses:
        '200':
          description: Current balance
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BalanceResponse'
              examples:
                sample:
                  value:
                    customer_id: cust-1
                    active_points: 100
                    pending_points: 50
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /v1/customers/{customerId}/transactions:
    get:
      tags:
        - Customers
      summary: Get customer transaction history
      operationId: getCustomerTransactions
      parameters:
        - $ref: '#/components/parameters/CustomerIdPath'
      responses:
        '200':
          description: Customer transaction list. Returns an empty array if the customer has no transactions.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TransactionResponse'
              examples:
                sample:
                  value:
                    - id: 1
                      customer_id: cust-1
                      order_id: order-1001
                      reference: null
                      return_id: null
                      kind: accrue
                      status: pending
                      points: 170
                      home24_merch_cents: 1200
                      mirakl_merch_cents: 1000
                      order_total_cents: 2200
                      shipping_cents: 490
                      currency: EUR
                      occurred_at: "2026-04-16T18:00:00Z"
                    - id: 2
                      customer_id: cust-1
                      order_id: order-1001
                      reference: null
                      return_id: null
                      kind: confirm
                      status: active
                      points: 170
                      home24_merch_cents: null
                      mirakl_merch_cents: null
                      order_total_cents: null
                      shipping_cents: null
                      currency: null
                      occurred_at: "2026-04-16T18:05:00Z"
                empty:
                  value: []
        '500':
          $ref: '#/components/responses/InternalServerError'

components:
  parameters:
    CustomerIdPath:
      name: customerId
      in: path
      required: true
      description: Customer identifier
      schema:
        type: string

    IdempotencyKeyHeader:
      name: Idempotency-Key
      in: header
      required: false
      description: Optional idempotency key header. Used when idempotency_key is not provided in the JSON body. The current implementation uses this key for duplicate protection; repeated or conflicting requests return 409 Conflict and are not replayed as 200 OK responses.
      schema:
        type: string

  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            badRequest:
              value:
                error: invalid input

    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            notFound:
              value:
                error: not found

    Conflict:
      description: Conflict with current state
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            conflict:
              value:
                error: conflict

    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            internalError:
              value:
                error: internal server error

  schemas:
    StatusResponse:
      type: object
      properties:
        status:
          type: string
          enum:
            - created
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    AccrueRequest:
      type: object
      description: Request for creating pending loyalty points from an order. Points are calculated by the service from the merchandise amounts. home24 merchandise earns 10 points per EUR, Mirakl/marketplace merchandise earns 5 points per EUR, shipping earns 0 points, and only EUR is accepted.
      properties:
        customer_id:
          type: string
        order_id:
          type: string
        home24_merch_cents:
          type: integer
          minimum: 0
          description: Merchandise amount in cents for home24 items.
        mirakl_merch_cents:
          type: integer
          minimum: 0
          description: Merchandise amount in cents for Mirakl/marketplace items.
        shipping_cents:
          type: integer
          minimum: 0
          description: Shipping amount in cents. Shipping is stored for traceability but does not earn points.
        currency:
          type: string
          enum:
            - EUR
        idempotency_key:
          type: string
      required:
        - customer_id
        - order_id
        - home24_merch_cents
        - mirakl_merch_cents
        - shipping_cents
        - currency

    ConfirmRequest:
      type: object
      properties:
        customer_id:
          type: string
        order_id:
          type: string
        idempotency_key:
          type: string
      required:
        - customer_id
        - order_id

    RevokeRequest:
      type: object
      properties:
        customer_id:
          type: string
        order_id:
          type: string
        return_id:
          type: string
        idempotency_key:
          type: string
      required:
        - customer_id
        - order_id
        - return_id

    RedeemRequest:
      type: object
      properties:
        customer_id:
          type: string
        reference:
          type: string
        points:
          type: integer
        idempotency_key:
          type: string
      required:
        - customer_id
        - reference
        - points

    RestoreRequest:
      type: object
      properties:
        customer_id:
          type: string
        reference:
          type: string
        idempotency_key:
          type: string
      required:
        - customer_id
        - reference

    BalanceResponse:
      type: object
      properties:
        customer_id:
          type: string
        active_points:
          type: integer
        pending_points:
          type: integer
      required:
        - customer_id
        - active_points
        - pending_points

    TransactionResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          minimum: 0
        customer_id:
          type: string
        order_id:
          type: string
          nullable: true
        reference:
          type: string
          nullable: true
        return_id:
          type: string
          nullable: true
        kind:
          type: string
          description: Type of ledger entry.
          enum:
            - accrue
            - confirm
            - revoke
            - redeem
            - restore
        status:
          type: string
          description: Ledger status. pending is used for accrued points waiting for confirmation. active means that the ledger entry is valid and affects the balance or represents a completed action such as confirm, redeem, restore, or revoke.
          enum:
            - pending
            - active
        points:
          type: integer
        home24_merch_cents:
          type: integer
          nullable: true
          description: Merchandise amount in cents for home24 items. Present for accrue entries.
        mirakl_merch_cents:
          type: integer
          nullable: true
          description: Merchandise amount in cents for Mirakl/marketplace items. Present for accrue entries.
        order_total_cents:
          type: integer
          nullable: true
          description: Sum of home24_merch_cents and mirakl_merch_cents. Shipping is not included.
        shipping_cents:
          type: integer
          nullable: true
          description: Shipping amount in cents. Shipping does not earn points.
        currency:
          type: string
          nullable: true
          description: Currency used for the order amount. Currently only EUR is accepted for accrual.
        occurred_at:
          type: string
          format: date-time
      required:
        - id
        - customer_id
        - kind
        - status
        - points
        - occurred_at